# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WowoClean API",
        "description": "API untuk sistem manajemen kontainer limbah B3 WowoClean",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login dan dapatkan JWT token",
                "operationId": "da2ca4d7a5cdec77b8b45a20373535cb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "admin@wowoclean.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login berhasil, token dikembalikan"
                    },
                    "401": {
                        "description": "Kredensial salah"
                    }
                }
            }
        },
        "/api/v1/profile": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Lihat profil user yang sedang login",
                "operationId": "e7fa93473c44547baf6183ba42322df2",
                "responses": {
                    "200": {
                        "description": "Data profil user"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout dan invalidate token",
                "operationId": "c4eddaf1a2e990cc90e4326c0d731d41",
                "responses": {
                    "200": {
                        "description": "Logout berhasil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/gateway/containers": {
            "get": {
                "tags": [
                    "Gateway - Containers"
                ],
                "summary": "Ambil semua data kontainer via Gateway (admin & user)",
                "operationId": "831b0aceaf0baf8e3721d25fa92ec8d0",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Chemical"
                        }
                    },
                    {
                        "name": "min_weight",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "example": 500
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar kontainer berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Gateway - Containers"
                ],
                "summary": "Tambah kontainer baru via Gateway (admin only)",
                "operationId": "383c17dd8f0d106fd042947e8d9ae24d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "container_id",
                                    "waste_type",
                                    "weight_kg",
                                    "status"
                                ],
                                "properties": {
                                    "container_id": {
                                        "type": "string",
                                        "example": "XY12345"
                                    },
                                    "waste_type": {
                                        "type": "string",
                                        "example": "Chemical"
                                    },
                                    "weight_kg": {
                                        "type": "number",
                                        "example": 500
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "Active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Kontainer berhasil dibuat"
                    },
                    "403": {
                        "description": "Forbidden - bukan admin"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/gateway/containers/{id}/archive": {
            "patch": {
                "tags": [
                    "Gateway - Containers"
                ],
                "summary": "Arsipkan kontainer via Gateway (admin only)",
                "operationId": "f6210060bd215f110527c05a61518d87",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil diarsipkan"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/gateway/containers/{id}": {
            "delete": {
                "tags": [
                    "Gateway - Containers"
                ],
                "summary": "Hapus kontainer via Gateway (admin only)",
                "operationId": "04a292a052141f6eae9f35e817937171",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil dihapus"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Gateway - Containers",
            "description": "Gateway - Containers"
        }
    ]
}